% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cal_defaults}
\alias{cal_defaults}
\title{Get default settings for calibration curves}
\usage{
cal_defaults(x = NULL)
}
\arguments{
\item{x}{ignored}
}
\value{
a list containing default arguments to supply to \code{pmcalibration::pmcalibration}
}
\description{
Get default settings for calibration curves
}
\keyword{internal}
