% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_stability}
\alias{get_stability}
\title{Get stability from internal_validate or internal_boot object}
\usage{
get_stability(x)
}
\arguments{
\item{x}{a internal_validate (see \code{\link{validate}}) or internal_boot
(see \code{\link{boot_optimism}}) object. The former should have been
created with method = "boot_optimism" or "boot_simple".}
}
\value{
a list containing the stability matrix (an n x B+1 matrix.
Each column contains predictions for p observations. First column contains
predictions from original model. Other columns are from boostrap models) and
the original binary outcome (y).
}
\description{
Get stability from internal_validate or internal_boot object
}
\keyword{internal}
