% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_plot.R
\name{mape_stability}
\alias{mape_stability}
\title{Mean absolute predictor error (MAPE) stability across bootstrap replicates}
\usage{
mape_stability(x)
}
\arguments{
\item{x}{an object produced by \code{\link{validate}} with method = "boot_\*" (or \code{\link{boot_optimism}} with method="boot")}
}
\value{
plots calibration (in)stability.
Invisibly returns a list containing data for each curve (p=x-axis, pc=y-axis).
The first element of this list is the apparent curve (original model on original outcome).
}
\description{
A MAPE (in)stability plot shows mean absolute predictor error (average absolute difference between original
predicted risk and predicted risk from B bootstrap models) as a function of apparent
predicted risk (prediction from original/development model). See Riley and Collins (2023).
}
\examples{
library(pminternal)
set.seed(456)
# simulate data with two predictors that interact
dat <- pmcalibration::sim_dat(N = 2000, a1 = -2, a3 = -.3)
mean(dat$y)
dat$LP <- NULL # remove linear predictor

# fit a (misspecified) logistic regression model
m1 <- glm(y ~ ., data=dat, family="binomial")

# internal validation of m1 via bootstrap optimism with 20 resamples
# B = 20 for example but should be >= 200 in practice
m1_iv <- validate(m1, method="boot_optimism", B=20)

mape_stability(m1_iv)

}
\references{
Riley RD, Collins GS. (2023). Stability of clinical prediction models developed using statistical or machine learning methods. Biom J. doi:10.1002/bimj.202200302. Epub ahead of print.
}
