% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pminternal-package.R
\docType{package}
\name{pminternal-package}
\alias{pminternal}
\alias{pminternal-package}
\title{pminternal: Internal Validation of Clinical Prediction Models}
\description{
Conduct internal validation of a clinical prediction model for a binary outcome. Produce bias corrected performance metrics (c-statistic, Brier score, calibration intercept/slope) via bootstrap (simple bootstrap, bootstrap optimism, .632 optimism) and cross-validation (CV optimism, CV average). Also includes functions to assess model stability via bootstrap resampling. See Steyerberg et al. (2001) \doi{10.1016/s0895-4356(01)00341-9}; Harrell (2015) \doi{10.1007/978-3-319-19425-7}; Riley and Collins (2023) \doi{10.1002/bimj.202200302}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stephenrho/pminternal}
  \item \url{https://stephenrho.github.io/pminternal/}
  \item Report bugs at \url{https://github.com/stephenrho/pminternal/issues}
}

}
\author{
\strong{Maintainer}: Stephen Rhodes \email{steverho89@gmail.com} [copyright holder]

}
\keyword{internal}
