% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_optimism.R
\name{print.internal_boot}
\alias{print.internal_boot}
\title{Print a internal_boot object}
\usage{
\method{print}{internal_boot}(x, digits = 2, ...)
}
\arguments{
\item{x}{an object created with \code{boot_optimism}}

\item{digits}{number of digits to print (default = 2)}

\item{...}{additional arguments to print}
}
\value{
invisibly returns \code{x} and prints estimates to console
}
\description{
Print a internal_boot object
}
