% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{print.internal_cv}
\alias{print.internal_cv}
\title{Print a internal_cv object}
\usage{
\method{print}{internal_cv}(x, digits = 2, ...)
}
\arguments{
\item{x}{an object created with \code{crossval}}

\item{digits}{number of digits to print (default = 2)}

\item{...}{additional arguments to print}
}
\value{
invisibly returns \code{x} and prints estimates to console
}
\description{
Print a internal_cv object
}
