% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{print.internal_validate}
\alias{print.internal_validate}
\title{print a internal_validate object}
\usage{
\method{print}{internal_validate}(x, digits = 2, ...)
}
\arguments{
\item{x}{a \code{internal_validate} object}

\item{digits}{number of digits to print}

\item{...}{optional arguments passed to print}
}
\value{
prints a summary
}
\description{
print a internal_validate object
}
