% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{print.internal_validatesummary}
\alias{print.internal_validatesummary}
\title{Print summary of internal_validate object}
\usage{
\method{print}{internal_validatesummary}(x, digits = 2, ...)
}
\arguments{
\item{x}{a \code{internal_validatesummary} object}

\item{digits}{number of digits to print}

\item{...}{ignored}
}
\value{
invisible(x) - prints a summary
}
\description{
Print summary of internal_validate object
}
