% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_plot.R
\name{dcurve_stability}
\alias{dcurve_stability}
\title{Plot decision curve stability across bootstrap replicates}
\usage{
dcurve_stability(x, thresholds = seq(0, 0.99, by = 0.01))
}
\arguments{
\item{x}{an object produced by \code{\link{validate}} with method = "boot_\*" (or \code{\link{boot_optimism}} with method="boot")}

\item{thresholds}{points at which to evaluate the decision curves (see \code{dcurves::dca})}
}
\value{
plots decision curve (in)stability.
Invisibly returns a list containing data for each curve. These are returned from \code{dcurves::dca}.
The first element of this list is the apparent curve (original model on original outcome).
}
\description{
A decision curve (in)stability plot shows decision curves for bootstrap
models evaluated on original outcome. A stable model should produce
curves that differ minimally from the 'apparent' curve.
See Riley and Collins (2023).
}
\examples{
set.seed(456)
# simulate data with two predictors that interact
dat <- pmcalibration::sim_dat(N = 2000, a1 = -2, a3 = -.3)
mean(dat$y)
dat$LP <- NULL # remove linear predictor

# fit a (misspecified) logistic regression model
m1 <- glm(y ~ ., data=dat, family="binomial")

# internal validation of m1 via bootstrap optimism with 20 resamples
# B = 20 for example but should be >= 200 in practice
m1_iv <- validate(m1, method="boot_optimism", B=20)

dcurve_stability(m1_iv)

}
\references{
Riley RD, Collins GS. (2023). Stability of clinical prediction models developed using statistical or machine learning methods. Biom J. doi:10.1002/bimj.202200302. Epub ahead of print.
}
